+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 - Added more helpful error messages for the Firebase Hosting GitHub Action (#5749)
 - Upgrade Firestore emulator to 1.17.4
+- Web Frameworks will no longer try to deploy unsupported versions of NodeJS to Cloud Functions (#5733)